pyproject.toml
[project]
name = "ytdlp-tray"
version = "1.0.0"
description = "System-tray yt-dlp download server for YouTube"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0",
    "flask-cors>=4.0",
    "pystray>=0.19",
    "pillow>=10.0",
    "pyyaml>=6.0",
]

[project.scripts]
ytdlp-tray = "ytdlp_tray.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ytdlp_tray"]

[tool.pyinstaller]
# Used as reference — actual build command is in build.ps1
name = "ytdlp-tray"
onefile = true
windowed = true
icon = "assets/icon.ico"